Re-organize tests to avoid duplication
